\name{gtheoryr-package}
\alias{gtheoryr-package}
\alias{gtheoryr}
\title{gtheoryr: Simple Generalizability Theory for R}
\description{
Small, beginner-friendly helpers for estimating variance components in simple
generalizability theory designs. The package currently supports:
\itemize{
  \item a crossed persons-by-items design via \code{gstudy_pxi()}
  \item a nested items-within-person design via \code{gstudy_nested_ip()}
  \item simple D-studies via \code{dstudy_pxi()} and \code{dstudy_nested_ip()}
}
}
\docType{package}
